code/backend/backend/requirements.txt
asgiref==3.9.1
astroid==3.3.11
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
black==25.1.0
certifi==2025.7.14
cffi==1.17.1
channels==4.2.2
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
constantly==23.10.4
cryptography==45.0.5
daphne==4.2.1
dill==0.4.0
Django==5.1.7
django-channels==0.7.0
django-cors-headers==4.7.0
django-rest-passwordreset==1.5.0
django-taggit==6.1.0
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.1
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.7.1
flake8==7.3.0
hyperlink==21.0.0
idna==3.10
inflection==0.5.1
isort==6.0.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
mccabe==0.7.0
mypy-extensions==1.1.0
oauthlib==3.3.1
packaging==25.0
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.3.8
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycodestyle==2.14.0
pycparser==2.22
pyflakes==3.4.0
PyJWT==2.10.1
pylint==3.3.7
pyOpenSSL==25.1.0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
requests-oauthlib==2.0.0
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
tomli==2.2.1
tomlkit==0.13.3
txaio==23.6.1
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
zope.interface==7.2
psycopg2-binary
python-decouple
